+++ hub-service/src/main/java/com/sparta/lucky/hub/application/RouteService.java
@@ -6,6 +6,8 @@ import com.sparta.lucky.hub.common.exception.HubErrorCode;
 import com.sparta.lucky.hub.domain.HubRoute;
 import com.sparta.lucky.hub.infrastructure.HubRouteRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,33 +20,20 @@ public class RouteService {
 
     private final HubService hubService;
     private final HubRouteRepository hubRouteRepository;
-    private final HubToHubPathFinder hubToHubPathFinder;
 
-    @Transactional(readOnly = true)
-    public GetRouteResult getRoute(UUID originHubId, UUID destinationHubId) {
-
-        // 출발 허브 == 도착 허브인 경우
-        if (originHubId.equals(destinationHubId)) {
-            return GetRouteResult.of(originHubId, destinationHubId, 0, 0, List.of(originHubId));
-        }
-
-        // Dijkstra로 최단 경로 탐색
-        List<HubRoute> routes = hubRouteRepository.findAllByDeletedAtIsNull();
-        HubToHubPathFinder.PathResult result = hubToHubPathFinder.findShortestPath(routes, originHubId, destinationHubId);
-
-        return GetRouteResult.of(originHubId, destinationHubId, result.totalDuration(), result.totalDistance(), result.path());
-    }
-
+    @Cacheable(cacheNames = "routes", key = "'all'")
     @Transactional(readOnly = true)
     public List<HubRoute> getHubRoutes() {
         return hubRouteRepository.findAllByDeletedAtIsNull();
     }
 
+    @CacheEvict(cacheNames = "routes", key = "'all'")
     @Transactional
     public void saveHubRoute(HubRoute hubRoute) {
         hubRouteRepository.save(hubRoute);
     }
 
+    @CacheEvict(cacheNames = "routes", key = "'all'")
     @Transactional
     public HubRoute createRoute(UUID originHubId, UUID destinationHubId, int distance, int duration) {
         hubService.getHub(originHubId);
@@ -53,6 +42,7 @@ public class RouteService {
         return hubRouteRepository.save(route);
     }
 
+    @CacheEvict(cacheNames = "routes", key = "'all'")
     @Transactional
     public HubRoute updateRoute(UUID routeId, int distance, int duration) {
         HubRoute route = findActiveRoute(routeId);
@@ -60,6 +50,7 @@ public class RouteService {
         return route;
     }
 
+    @CacheEvict(cacheNames = "routes", key = "'all'")
     @Transactional
     public void deleteRoute(UUID routeId, UUID deletedBy) {
         HubRoute route = findActiveRoute(routeId);
